Reworked the data types to int. Added correct HP calc. Added correct AC calc.

+++ src/main/Die.java
@@ -3,7 +3,7 @@ import java.util.Random;
 
 public class Die
 {
-  public static double rollDie(double nSides, double dieCount)
+  public static int rollDie(int nSides, int dieCount)
   {
 	int output = 0;
 	Random r = new Random();
@@ -12,16 +12,16 @@ public class Die
 	  // Weil nextInt(int) von 0 bis nSides-1 geht.
 	  output += r.nextInt((int)nSides) + 1;
 	}
-	return (double) output;
+	return (int) output;
   }
   
-  public static double rollDie_recursively(double nSides, double dieCount)
+  public static int rollDie_recursively(int nSides, int dieCount)
   {
 	int output = 0;
 	Random r = new Random();
 	output += r.nextInt((int)nSides) + 1;
-	dieCount -= 1.0;
-	output += (dieCount > 0) ? rollDie_recursively(nSides, dieCount) : 0.0;
+	dieCount -= 1;
+	output += (dieCount > 0) ? rollDie_recursively(nSides, dieCount) : 0;
 	return output;
   }
 }
